requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
celery==5.3.4
redis==4.6.0
python-telegram-bot==20.3
django-celery-results==2.5.1
django-celery-beat==2.5.0
